Seperate HTML Building Code from parsing code to improve readability, add ParsedLog as container class for the final log data, for future decoupling
