+++ LuckParser/LuckParser.csproj
@@ -78,6 +78,7 @@
     <Compile Include="Controllers\Controller1.cs" />
     <Compile Include="Controllers\FlashWindow.cs" />
     <Compile Include="Controllers\GW2APIController.cs" />
+    <Compile Include="Controllers\HTMLBuilder.cs" />
     <Compile Include="Controllers\HTMLHelper.cs" />
     <Compile Include="Controllers\ParseHelper.cs" />
     <Compile Include="MainForm.cs">
@@ -86,6 +87,7 @@
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
+    <Compile Include="Models\DataModels\ParsedLog.cs" />
     <Compile Include="Models\ParseModels\Players\AbstractMasterPlayer.cs" />
     <Compile Include="Models\ParseModels\Players\AbstractPlayer.cs" />
     <Compile Include="Models\ParseModels\Boons\BoonsGraphModel.cs" />
